Convert Fort Rox details
